# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/restaurantes/{restaurante}/avaliacoes": {
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Create a new review for a specific restaurant",
                "operationId": "66cbc9d4df3200401aaee2a1beb30c6c",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "description": "ID of the restaurant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "avaliacao"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "avaliacao": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.5
                                    },
                                    "comentario": {
                                        "type": "string",
                                        "example": "Great place!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "restaurante_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "avaliacao": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 4.5
                                        },
                                        "comentario": {
                                            "type": "string",
                                            "example": "Great place!"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-05-21T00:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-05-21T00:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid request data"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Restaurant not found"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "user_id": [
                                                    "The user id field is required."
                                                ]
                                            }
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurantes": {
            "get": {
                "tags": [
                    "Restaurant"
                ],
                "summary": "List all restaurants",
                "operationId": "5138987d5cae8d46f6127cc7d3613ce8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nome": {
                                                "type": "string"
                                            },
                                            "endereco": {
                                                "type": "string"
                                            },
                                            "cidade": {
                                                "type": "string"
                                            },
                                            "cep": {
                                                "type": "string"
                                            },
                                            "telefone": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "tipo_cozinha": {
                                                "type": "string"
                                            },
                                            "descricao": {
                                                "type": "string",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurant"
                ],
                "summary": "Create a new restaurant",
                "operationId": "2722e8c6d62c7164e697e9989aad970a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "endereco": {
                                        "type": "string"
                                    },
                                    "cidade": {
                                        "type": "string"
                                    },
                                    "cep": {
                                        "type": "string"
                                    },
                                    "telefone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "tipo_cozinha": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created restaurant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "type": "string"
                                        },
                                        "endereco": {
                                            "type": "string"
                                        },
                                        "cidade": {
                                            "type": "string"
                                        },
                                        "cep": {
                                            "type": "string"
                                        },
                                        "telefone": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "tipo_cozinha": {
                                            "type": "string"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Error creating restaurant"
                    }
                }
            }
        },
...
        "/restaurantes/{id}/avaliacoes": {
            "get": {
                "tags": [
                    "Restaurant"
                ],
                "summary": "List all the reviews of a restaurant",
                "operationId": "2b45ada20378a390f25202f8fba83416",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID restaurant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "restaurante_id": {
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "avaliacao": {
                                                "type": "number",
                                                "format": "float",
                                                "maximum": 5,
                                                "minimum": 1
                                            },
                                            "comentario": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    }
}